Got working echo with dfgm handler test in ipc_interface. interfaces -> tcp_interfaces.
