+++ dev_tools/cli_test_msg/src/main.rs
@@ -29,7 +29,7 @@ fn main() {
     let msg_time: u64 = timestamp_to_epoch(timestamp.clone()).unwrap();
     let msg_time_bytes = msg_time.to_le_bytes().to_vec();
 
-    let data: Msg = Msg::new(22,3,0,0,msg_time_bytes);
+    let data: Msg = Msg::new(22,3,25,25,msg_time_bytes);
 
     let mut stream = TcpStream::connect((Ipv4Addr::new(127, 0, 0, 1), port)).unwrap();
     let output_stream = &mut stream;
