+++ handlers/dfgm_handler/Cargo.toml
@@ -6,5 +6,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-interfaces = { path = "../../libs/interfaces" }
+tcp_interface = { path = "../../libs/tcp_interface" }
 common = {path = "../../libs/common"}
