+++ handlers/handler/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-interfaces = {path = "../../libs/interfaces" }
+tcp_interface = {path = "../../libs/tcp_interface" }
