+++ libs/tcp_interface/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "interfaces"
+name = "tcp_interface"
 version = "0.1.0"
 edition = "2021"
